--- Monika Donner, 19.03.2009 086.) Correction in SubR second:       Definition for integer :: itim       >> SubR SECOND 087.) Correction in SubR PipeSurfaceConnectionQs :       Formula:  flow = PS1 * (-1.0)  is corrected       >> SubR PipeSurfaceConnectionQs --- Monika Donner, 19.03.2009

+++ _import/trunk/fortran/RMA10S_TUHH/source/second.f90
@@ -1,9 +1,10 @@
+!     Last change:  MD   19 Mar 2009    1:02 pm
 subroutine second (ta)
 implicit none
 save
 ! this version is for microsoft fortran
 data itim /0/
-integer (kind = 2) :: itim
+integer :: itim
 real (kind = 8) , intent (out):: ta
 
 integer :: clock1, rate
